backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary>=2.9.9
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
yt-dlp>=2024.0.0
Pillow>=10.0.0
python-dotenv==1.0.0
websockets==12.0

